Also fixes an incorrect asv.

```
In [2]: ts = pd.Timestamp.now("US/Pacific")                                                                                                                                                                        

In [3]: %timeit ts.day_name()                                                                                                                                                                                      
9.13 µs ± 309 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- PR
26.4 µs ± 1.23 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)   # <-- master
```